+++ frontend/app/components/BootstrapClient.js
@@ -1,4 +1,4 @@
-"use client";
+'use client';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useEffect } from 'react';
